Get multiple github releases
Some repos create multiple different releases, not all releases hold all
artefacts. Therefore it is not guaranteed that the last release actually
holds the artefact we want.

Thus we pull multiple releases and find the newest artefact we can find.
We reley on github ordering the result by publishin time.